refactor(tests): consolidate preview2 tests into codegen
